商品左侧首页分类显示

+++ bilistore_parent/bilistore_manager/bilistore_manager_mapper/src/main/java/com/manager/mapper/TbItemCatMapper.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface TbItemCatMapper {
 
-    List<TbItemCat> getTbItemCatByParentId(@Param("parentId") Integer parentId);
+    List<TbItemCat> getTbItemCatByParentId(@Param("parentId") Long parentId);
 }
