+++ bilistore_parent/bilistore_manager/bilistore_manager_service/src/main/java/com/manager/service/TbItemCatService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface TbItemCatService {
 
-    List<EasyUI_Tree_ReturnType> getTbItemCatByParentId(Integer id);
+    List<EasyUI_Tree_ReturnType> getTbItemCatByParentId(Long id);
 }
